+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Bug fixes
 
+- Make `drake_plan(transform = slice())` understand `.id` and grouping variables (#963).
 - Repair `clean(garbage_collection = TRUE, destroy = TRUE)`. Previously it destroyed the cache before trying to collect garbage.
 
 ## New features